Create README.md
We use a neural network to map key words in development aid project descriptions to policy objectives. What we wanted to do is to predict whether a development aid contribution had a certain policy objective e.g. gender equality based on its description. A description of this project can be found here on my blog: